feat: adds format option with support for table and csv output & removes aws-sdk dependency
BREAKING CHANGE: removes the pretty parameter and replaces with format that accepts either string, json, table, or csv. adds a log parameter that can be used to suppress logging, which can be used in conjunction with the new .getEntries() method, for full control over the performance entries, which will return all native and custom entries of the types and in the format passed into the constructor.